feat: add delete and patch controller functions

+++ backend/controllers/workoutController.js
@@ -41,10 +41,15 @@ const createWorkout = async (req, res) => {
 // delete a workout
 const deleteWorkout = async (req, res) => {
     const {id} = req.params
+
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({error: 'Could not find workout'});
+    }
+
     const workout = await Workout.findByIdAndDelete({_id: id});
 
     if (!workout) {
-        return res.status(404).json({error: 'No such workout'});
+        return res.status(404).json({error: 'Could not find workout'});
     }
 
     res.status(200).json(workout);
@@ -52,11 +57,29 @@ const deleteWorkout = async (req, res) => {
 
 
 // update a workout
+const updateWorkout = async (req, res) => {
+    const {id} = req.params
+
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({error: 'Could not find workout'});
+    }
+
+    const workout = await Workout.findByIdAndUpdate({_id: id}, {
+        ...req.body
+    });
+
+    if (!workout) {
+        return res.status(404).json({error: 'Could not find workout'});
+    }
+
+    res.status(200).json(workout);
+}
 
 
 module.exports = {
     createWorkout,
     getWorkouts,
     getWorkout,
-    deleteWorkout
+    deleteWorkout,
+    updateWorkout
 }
